Update happiness_jar
This is an application that might serve as an digital version of Happiness Jar, a mind training tool that helps people enjoy and appreciate little things and moments in life :) It has two basic functionalities: firstly, adding new memories with assigned dates, and secondly, randomly choosing one memory from the jar. 
I'm a beginner. While preparing this project, I've learnt quite a lot:
1) Reading from and writing in a .txt file (the app creates and uses a .txt file instead of working only in RAM)
2) Thinking in terms of classes, not just in terms of variables and functions (object oriented programming, oh yeah)
3) Using "switches" - local variables, usually booleans that have one value at the beginning of a function or a loop and are updated inside the code under certain circumstances. The update lets me control the flow of the function/loop.
4) Last but not least - I've got a long way to go in terms of thinking like a coder :)