Create perm_check_codility
A non-empty zero-indexed array A consisting of N integers is given. A permutation is a sequence containing each element from 1 to N once, and only once. The goal is to check whether array A is a permutation.
Write a function: def solution(A), that, given a zero-indexed array A, returns 1 if array A is a permutation and 0 if it is not.
https://codility.com/programmers/task/perm_check/