Create missing_integer_codility
I attempted this exercise three times, with first attempt being a total disaster of 11% :) turns out that I didn't fully comprehend what this task was really about. Useful note: 
for A with 1 element possible outcomes are 1 (if A != 1) and 2 (if A == 1), e.g. 
A = [-200] -> the 1st positive number that's missing is 1, 
A = [345] -> the 1st positive number that's missing is still 1
so:
this task is not about arrays being just shuffled permutations with one number missing; it's about narrowing the number of elements I want to take into account :)